Update README.md
There was a case when a user went straight to the PyCharm to read the Notebook. That not only didn't resolve the file, but have also broken the meta-data in the file, so it wouldn't open in IntelliJ as Kotlin Notebook anymore. We will fix this bug, but meanwhile, it's better to move the comment about the Kotlin Notebooks Plugin higher because most users currently view the .ipynb file as a Python file.