Fix freezes on windows
When java process spawns another process via ProcessBuilder
the stdout and stderr of the just spawned process are piped by default
to the original java process. If no one reads these stdout/stderr
streams the underlying process can get blocked by "forever waiting"
for host process to empty the pipe buffers.

One of the solution is just discard stdout.
Because gradle profiler also writes it to the profile.log file.

However, stderr should be still read so this can be done in a separate thread.